removed custom auth class and extended userservice from defaultOAuth2UserService
