+++ API/src/main/java/com/grabit/app/config/SecurityConfig.java
@@ -7,17 +7,18 @@ import org.springframework.security.config.annotation.web.configuration.EnableWe
 import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
 import org.springframework.security.web.SecurityFilterChain;
 
-import com.grabit.app.service.CustomOAuth2UserService;
+import com.grabit.app.service.UserService;
+
 import org.springframework.security.web.authentication.www.BasicAuthenticationFilter;
 
 @EnableWebSecurity
 @Configuration
 public class SecurityConfig {
 
-    private final CustomOAuth2UserService customOAuth2UserService;
+    private final UserService userService;
 
-    public SecurityConfig(CustomOAuth2UserService customOAuth2UserService) {
-        this.customOAuth2UserService = customOAuth2UserService;
+    public SecurityConfig(UserService userService) {
+        this.userService = userService;
     }
 
     @Bean
@@ -29,7 +30,7 @@ public class SecurityConfig {
                 .addFilterBefore(new AuthFilter(), BasicAuthenticationFilter.class)
                 .oauth2Login(oauth2 -> oauth2
                         .userInfoEndpoint(userInfo -> userInfo
-                                .userService(customOAuth2UserService)))
+                                .userService(userService)))
                 .csrf(AbstractHttpConfigurer::disable);
         return http.build();
     }
