--- API/src/main/java/com/grabit/app/service/CustomOAuth2UserService.java
@@ -1,26 +1,0 @@
-package com.grabit.app.service;
-
-import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
-import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
-import org.springframework.security.oauth2.core.user.OAuth2User;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
-@Service
-public class CustomOAuth2UserService extends DefaultOAuth2UserService {
-
-    private final UserService userService;
-
-    public CustomOAuth2UserService(UserService userService) {
-        this.userService = userService;
-    }
-
-    @Override
-    @Transactional
-    public OAuth2User loadUser(OAuth2UserRequest userRequest) {
-        OAuth2User oAuth2User = super.loadUser(userRequest);
-        String githubID = oAuth2User.getAttribute("login").toString();
-        userService.saveOrUpdateUser(githubID);
-        return oAuth2User;
-    }
-}
