+++ API/src/main/java/com/grabit/app/service/UserService.java
@@ -2,6 +2,9 @@ package com.grabit.app.service;
 
 import java.time.LocalDate;
 
+import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
+import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
+import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -9,7 +12,7 @@ import com.grabit.app.model.User;
 import com.grabit.app.repository.UserRepository;
 
 @Service
-public class UserService {
+public class UserService extends DefaultOAuth2UserService {
     private final UserRepository userRepository;
 
     public UserService(UserRepository userRepository) {
@@ -28,5 +31,14 @@ public class UserService {
         }
 
         userRepository.save(user);
+    }
+
+    @Override
+    @Transactional
+    public OAuth2User loadUser(OAuth2UserRequest userRequest) {
+        OAuth2User oAuth2User = super.loadUser(userRequest);
+        String githubID = oAuth2User.getAttribute("login").toString();
+        this.saveOrUpdateUser(githubID);
+        return oAuth2User;
     }
 }
